fix: comprehensive bug fix, architecture cleanup, and performance optimization
Bug fixes:
- StatsViewModel: stale stats snapshot now refreshes on each collect
- FullTextSearch: eliminate unsafe !! operators (local variable null check)
- WheelViewModel: isSpinning resets on coroutine cancellation
- WheelScreen: eliminate NPE on dialog result (!! -> smart cast)
- PdfViewer: printStackTrace -> Log.e
- ReaderContent: fix paragraphSpacing float truncation
- CacheManager.ChapterMeta.toDomain: add bookId parameter (was always 0)

Architecture cleanup:
- Remove dead code: AppException.kt (unused sealed class + custom Result<T>)
- Remove dead code: DataManager.kt (DataCleaner, DataManager unused)
- Remove duplicate TextPaginator from BookLoader.kt
- Remove unused SettingsViewModel injection from FlowReaderApp.kt
- Remove unused imports and dead nav imports from FlowReaderApp.kt
- Remove duplicate getBookCount() from ChapterDao (already in BookDao)
- Remove unused singleBookPickerLauncher from LibraryScreen

Performance:
- Add date index on ReadingStatsEntity for date-only queries
- Fix thread safety in CacheManager.getCacheStats() (synchronized)
- Replace hardcoded 0.75f cache hit rate with real hit/miss tracking

Config:
- Remove deprecated configureondemand from gradle.properties

+++ app/src/main/java/com/flowreader/app/ui/screens/reader/components/ReaderContent.kt
@@ -98,7 +98,7 @@ fun ReaderContent(
                         ),
                         color = textColor,
                         modifier = Modifier
-                            .padding(bottom = settings.paragraphSpacing.toInt().dp)
+                            .padding(bottom = settings.paragraphSpacing.dp)
                             .pointerInput(Unit) {
                                 detectTapGestures(
                                     onLongPress = {
